Redesign the way totals is displayed and the navigation UI
Very much WIP, mostly just experimenting with some design ideas.

GoatCounter has always had a "path-first" interface, rather than a
"totals first" interface. I always much preferred it like that myself.
The "Totals" at the top always seemed kind of misplaced to me, but I
wasn't sure where/how to do it better.

Placing it at the bottom under a separate header actually seems like a
pretty good solution. My original thought was to display it as a
half-width together with the "browsers", "systems", etc. charts (and
bring back the "top refs"), but that didn't really look too good.

Not 100% sold on putting it wider than the other charts yet. I kind of
like it so that it's more distinguished though.

Since there is now more space between the charts, the independent
scaling between the regular pages and Totals looks less weird, so this
fixes #291.

I've also removed the configurable scaling for now; I don't really like
how it works, and it clutters up the UI a bit. I think I'll replace it
with a "↕️" link to set the scaling to the max of a particular chart (or
something like that?)
TODO: do this in master

Some people may still prefer to have the Totals at the top, the
configurable dashboard discussed in #302 will enable that eventually
(should be fairly soon). You can also set the pagination for the paths
to e.g. 5 so it's still in view on page load.

This also ties in to the navigation, which is now a bit more consistent.
TODO: do in master
TODO: fix this up as it's a bit ugly now, but putting everything in the
top is probably the right call here.
TODO: implement the "save view", probably in a future PR, but best to
account for design now.
TODO: need to bring back "go" button, although I think we can just
reload on date select.