Put totals at the bottom
GoatCounter has always had a "path-first" interface, rather than a
"totals first" interface. I always much preferred it like that myself.
The "Totals" at the top always seemed kind of misplaced to me, but I
wasn't sure where/how to do it better.

Placing it at the bottom under a separate header actually seems like a
pretty good solution. My original thought was to display it as a
half-width together with the "browsers", "systems", etc. charts (and
bring back the "top refs"), but that didn't really look too good.

Some people may still prefer to have the Totals at the top, the
configurable dashboard discussed in #302 will enable that eventually
(should be fairly soon). You can also set the pagination for the paths
to e.g. 5 so it's still in view on page load.

The "Top referrals" is now also back as a separate chart at the bottom;
I found that actually worked better

TODO:

- Different scaling for "totals" and the paths.